Widen Node::Leaf/Branch's value-sum accumulator from f32 to f64
Node::Leaf/Branch's `w` is a running sum over up to `n` per-visit
values. f32 has only ~7 significant decimal digits (24-bit mantissa,
exact integers only up to 2^24), so a heavily-visited node in a
long-running search - plausibly the root, or any node near it - can
silently stop accumulating small increments well within reach of
real iteration counts, systematically biasing its reported average.

Per-visit values (rollout()/custom_evaluation()/GameResult::value())
stay f32 - there's no precision concern for a single bounded value,
only for summing many of them. Info.q and ply()'s reported average
both stay f32 too (cast down at the boundary); this is purely an
internal storage/arithmetic change, no public API is affected.

+++ arbor/src/lib.rs
@@ -53,10 +53,16 @@ pub trait GameState<P: Player, A: Action>: Copy + Debug + Display {
 enum Node<P: Player, A: Action> {
     //sibling?, action, player, value, visits, child
     //s,a,p,w,n,c
+    //
+    //Leaf/Branch's `w` is a running sum over up to `n` per-visit values (each in [0,1]); it's
+    //f64 rather than f32 specifically to avoid precision loss at high visit counts (f32 only
+    //has ~7 significant digits, so a heavily-visited node's sum can silently stop accumulating
+    //small increments well within the iteration counts a long-running search can reach).
+    //Terminal's `w` is a single proven value, never summed further, so f32 is precise enough.
     Unknown(bool,A),
     Terminal(bool,A,P,f32),
-    Leaf(bool,A,P,f32,u32),
-    Branch(bool,A,P,f32,u32,usize),
+    Leaf(bool,A,P,f64,u32),
+    Branch(bool,A,P,f64,u32,usize),
     Transpose(bool,A,usize),
 }
